backend/pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "1.0.0"
description = "RAG Chatbot Backend - FastAPI service with vector database and LLM integration"
authors = [
    {name = "RAG Chatbot Team", email = "team@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "structlog>=23.2.0",
    "python-multipart>=0.0.9",
    "pdfminer.six>=20221105",
    "python-docx>=1.0.0",
    "rapidfuzz>=3.6.1",
    "google-cloud-storage>=2.18.0",
    "httpx>=0.27.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "tiktoken>=0.5.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "lxml>=4.9.0",
    "urllib3>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

